+++ src/Application/Exception/Unavailability/TimeOutOfRangeException.php
@@ -2,21 +2,19 @@
 
 namespace Proximum\Vimeet\Application\Exception\Unavailability;
 
-use Proximum\Vimeet\Domain\Model\Event\Day;
-use Proximum\Vimeet\Domain\Time\TimeRangeView;
+use Proximum\Vimeet\Domain\Time\TimeRangeInterface;
 
 class TimeOutOfRangeException extends UnavailabilityException
 {
     public const BEGIN = 'begin';
     public const END   = 'end';
 
-    /** @var TimeRangeView */
-    public $day;
+    public TimeRangeInterface $day;
 
     /** @var string */
     public $period;
 
-    public function __construct(TimeRangeView $day, $period)
+    public function __construct(TimeRangeInterface $day, $period)
     {
         parent::__construct('Time selected is out of range');
         $this->day = $day;