+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Hotfix - PHP runtime error when unavailability is out of an event day
 
 ## [2.92.1] - 2021-05-18
 - Hotfix - Allow empty sheet in happening participants export